V1/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cma==4.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.2.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
evaluate==0.4.3
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.9.0
graphviz==0.21
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lm_eval==0.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
ml_collections==1.0.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.3
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pybind11==2.13.6
Pygments==2.19.1
pyparsing==3.2.3
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.3.1
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.47.1
triton==2.3.1
typepy==1.3.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
word2number==1.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
